requirements.txt
pyscreenshot
Pillow==6.2.1